Port test infra from illinker (#25)
This ports all of the test infrastructure from the linker repo. I intentionally removed any reflection pattern validation (since we don't want it in the linker either and it would not be possible anyway).

Obviously all of the various command line arguments for the linker don't work either (the attributes are there, but they do nothing).

Otherwise it works exactly like in the linker repo. As we implement features we should validate that the test infra can actually test them (it's very possible there are holes in the linker test infra).

As for the tests: I ported the existing test over as the "First".
I also copied one of the linker tests "UnusedMethodGetsRemoved" as-is... and it passes :wink:

Note: This is obviously a HUGE change - so I don't expect people review it really. What I did:
- Copied over the necessary files
- Adapted everything to XUnit (this is actually most of the work) - I had to use FluentAssertions as well, since XUnit is very opinionated (and stubborn) in some cases.
  - This has some effect on how tests are viewed in the Test Explorer - XUnit doesn't have the same ability as NUnit to programaticaly generate test cases - it can do it, but there's no tests name customization. I tried to keep the names reasonable.
- Fixed lot of null annotations - I tried to keep everything with nullable on, but wasn't able to do all of it - but most of it is nullable enabled.
- Rewrite the argument construction - currently mostly empty
- Removed any logging and reflection pattern validation - as the "product" doesn't have these capabilities yet