Add managed type system (#28)
This replaces the stub EcmaModule with the "real" EcmaModule from the type system.

The type system will provide services for us around resolving things to definitions and having a way to represent types/methods/fields in an interned universe.

For now, this doesn't do much, but notice that the EcmaModule we're now using is more powerful. It e.g. provides a `GetObject` API that allows resolving EntityHandles to their type system representation.