Track virtual method usage
This adds tracking virtual methods to ensure that if a type was allocated and provides an implementation to a virtual method that was used, we will generate the implementation as well.

Introduced a new node type that represents a constructed type, and another node type that represents a virtual method use. The constructed type node conditionally depends on virtual method implementations, conditioned on the virtual method being used.