Implement analyzing and writing manifest resources
Add a node for manifest resource (it's a handle after all).
Implement analysis and writing for it.
Add ability to token nodes to decide to not write themselves to the output (they can bring in other dependencies still).
Use this mechanism to special case linker descriptor resource, which is processed as a normal resource node, but it doesn't write itself to the output.

Added simple tests.